Add brisk_system and feature to wait for RPC to start
Hold thread references to be able to cancel them as part
of the future funcitonality.
Implement a waiting mechanisms to check/create for
brisk_system keyspace once Cassandra is ready to listen to
connections.
Since we run inside the same JVM, parhaps there is a better
way to use it besides connection to localhost.